Moved removeCourse method to SearchResponse as it is called when on the search on the frontEnd. Now open up test schedule that was saved at open of program
